Adding a folder
Adding a folder that includes all of my scripts and modified files. There is an index file that discusses what each of these is, which should be informational! Index includes a list of tasks Jacob needs to complete in order to transform the things in this file pile into an actually updated version of the platform.