Setup our flask application
Setup our Flask application which will run on AWS. This will allow us
to on demand generate analysis as well as automate analysis if we so please.
This will need to be integrated with the work in `ErgoAnalytics` so that
we can query data and run analysis as well as send the results to
infinity-gauntlet.

There were also a few changes and fixes made in ErgoAnalytics to get
it to run.