Query data from aws elasticsearch
With this change we will be able to query data from aws elasticsearch.
This will allow us to run our analysis on AWS on demand using real data.
Something to note that there is a known bug with python and urllib
that causes this to error out on macOSX, in order to get around this
you need to set the environment variable `NO_PROXY=*` in order to run
the dramatiq worker without causing a segmentation fault.