Add logging and fix ergoReport
ErgoReport was relying on old code and had not been updated. It was
using the experiments class which was replaced by the structuredData
classes. This should update the code to use the correct constructs.

It also added logging so we can hopefully better track what is happening
on the background job.
`