Create automated analysis
Created a cron dramatiq task using periodic to run automated analysis
for all wearables we have marked as automated. If there are no wearables
marked then we do not have any analysis to run. If there are wearables
then we will loop of the list and enqeueue our normal `safety_score_analysis`
job to run the analysis for that given job.

Also made a change to use `start_time` and `end_time` instead of
`from_date` and `till_date`. They are techncially time stamps so the
`date` was misleading. Also from and till is slightly more complicated
to think about then start and end. I was getting slightly confused
so I decided to simplify the names.