Fix loading from elastic search
Previously this code was assuming there was no timestamp in the values
we got from the wearable. This caused us to add timestamps twice
which was causing us to remove most data because it didn't fit the
right structure. With this change we will use the timestamp from elasticsearch
and remove the timestamp from the wearable.

In the future if we remove the timestamp from the wearable then
we can slightly modify this code so that it doesn't have to remove
the data from the hardware received wearables.