Handle no timestamp in raw data
If there is no timestamp in the raw data we get from elasticsearch
then we should just be able to get the raw value string from
elasticsearch and append the timestamp + a comma to the beginning of it.

The value formatter should handle the different formats of the data,
we shouldn't have to do any operations on the raw data. Splitting it
will be unnecessary because previously we were just splitting it
to put it back together without the timestamp. This change will
make it so that we just append the timestamp to the beginning of this
value string.