Move params to the scan call
We have a bit of code `params(preserve_order=True)` that we are using
in load elastic search class. We need this to make sure that the values
are loaded from elastic search, sorted, and when we scan they come in
the right order. Without this we will have unordered values due to the
nature of the scan method. However, the params call can not be used
with the `count` method call we use to check size. Therefore we must
move it from the query to the scan call in order to fix this issue.