Fix Gunicorn 20 error
This fixes two errors, one where we were not importing HTTPerror
appopriately in tasks.py which led to the app crashing and a second
error introduced by upgrading gunicorn. Gunicorn was upgraded to version
20 from 19.9 which was a major version upgrade and introduced a breaking
change. The change is documented by this issue on the git repo https://github.com/benoitc/gunicorn/issues/2159
it is related to removing an eval statement so we can no longer have
gunicorn call a function to create our app. We have gone with a solution
introduced in this issue, to call the create_app function and then
export a variable where this application is saved. We then call
this variable in our Procfile where we run our server using Gunicorn.

This should fix our issue for now, and if gunicorn fixes the issue
in the libarary then we could always revert but I think this should
work going forward.